Create Homework 9 - More on Classes - DayOfYear MODIFICATION
3. Day of the Year Modiﬁcation
Modify the DayOfYear class, written in an earlier Programming Challenge, to add a
constructor that takes two parameters: a string representing a month and an integer in the
range 0 through 31 representing the day of the month. The constructor should then initialize
the integer member of the class to represent the day speciﬁed by the month and day of month
parameters. The constructor should terminate the program with an appropriate error
message if the number entered for a day is outside the range of days for the month given.
Add the following overloaded operators:
++ preﬁx and postﬁx increment operators. These operators should modify the DayOfYear
object so that it represents the next day. If the day is already the end of the year, the
new value of the object will represent the ﬁrst day of the year.
--  preﬁx and postﬁx decrement operators. These operators should modify the DayOfYear
object so that it represents the previous day. If the day is already the ﬁrst day of the
year, the new value of the object will represent the last day of the year.